requirements.txt
###### Requirements without Version Specifiers ######`
requests
paho-mqtt
colorama
